feat: site use theme-generator
