--- src/layouts/components/Setting/RadioColor.module.less
@@ -1,17 +1,0 @@
-.panel {
-  display: flex;
-  justify-content: space-between;
-}
-
-.box {
-  cursor: pointer;
-  border: 2px solid;
-  padding: 4px;
-  border-radius: 50%;
-}
-
-.item {
-  width: 24px;
-  height: 24px;
-  border-radius: 50%;
-}
