+++ src/layouts/components/Setting/RadioRect.module.less
@@ -4,14 +4,16 @@
 }
 
 .rectItem {
-  border: 2px solid #e3e6eb;
-  padding: 8px;
+  border: none;
+  padding: 0;
   border-radius: 3px;
   cursor: pointer;
-}
-
-.rectItemSelected {
-  border-color: var(--td-brand-color);
+  position: relative;
+  .pickedItem {
+    position: absolute;
+    right: 0;
+    bottom: 6px;
+  }
 }
 
 .rectText {
