--- src/layouts/components/Setting/index.module.less
@@ -1,13 +1,0 @@
-.settingTitle {
-  margin: 32px 0 24px 0;
-  color: var(--td-text-color-primary);
-  font-size: 14px;
-  line-height: 22px;
-  font-weight: 500;
-}
-
-.settingSubTitle {
-  font-size: 14px;
-  line-height: 21px;
-  margin-bottom: 24px;
-}
