+++ src/layouts/index.tsx
@@ -1,8 +1,8 @@
 import React, { memo, useEffect } from 'react';
-import { Drawer, Layout } from 'tdesign-react';
+import { Layout } from 'tdesign-react';
 import throttle from 'lodash/throttle';
 import { useAppSelector, useAppDispatch } from 'modules/store';
-import { selectGlobal, toggleSetting, toggleMenu, ELayout, switchTheme } from 'modules/global';
+import { selectGlobal, toggleMenu, ELayout, switchTheme } from 'modules/global';
 import Setting from './components/Setting';
 import AppLayout from './components/AppLayout';
 import Style from './index.module.less';
@@ -31,16 +31,7 @@ export default memo(() => {
   return (
     <Layout className={Style.panel}>
       <AppContainer />
-      <Drawer
-        destroyOnClose
-        visible={globalState.setting}
-        size='458px'
-        footer={false}
-        header='页面配置'
-        onClose={() => dispatch(toggleSetting())}
-      >
-        <Setting />
-      </Drawer>
+      <Setting />
     </Layout>
   );
 });
