+++ src/pages/Login/components/Header/index.tsx
@@ -1,25 +1,17 @@
 import React from 'react';
 import { Button } from 'tdesign-react';
-import { LogoGithubIcon, HelpCircleIcon, SettingIcon } from 'tdesign-icons-react';
-import { useAppDispatch } from 'modules/store';
-import { toggleSetting } from 'modules/global';
+import { LogoGithubIcon, HelpCircleIcon } from 'tdesign-icons-react';
 
 import LogoFullIcon from 'assets/svg/assets-logo-full.svg?component';
 import Style from './index.module.less';
 
 export default function Header() {
-  const dispatch = useAppDispatch();
-
   const navToGitHub = () => {
     window.open('https://github.com/tencent/tdesign-react-starter');
   };
 
   const navToHelper = () => {
     window.open('http://tdesign.tencent.com/starter/docs/react/get-started');
-  };
-
-  const toggleSettingPanel = () => {
-    dispatch(toggleSetting());
   };
 
   return (
@@ -44,15 +36,6 @@ export default function Header() {
             onClick={navToHelper}
           >
             <HelpCircleIcon className={Style.icon} />
-          </Button>
-          <Button
-            className={Style.operationsButton}
-            theme='default'
-            shape='square'
-            variant='text'
-            onClick={toggleSettingPanel}
-          >
-            <SettingIcon className={Style.icon} />
           </Button>
         </div>
       </header>
